# docs/swagger.yaml
basePath: /api
definitions:
  api.JsonError:
    properties:
      message:
        example: error message
        type: string
    type: object
  service.CategoryData:
    properties:
      genre_id:
        type: integer
    required:
    - genre_id
    type: object
  service.Celebrity:
    properties:
      birth_date:
        example: "1969-12-28T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Linus Torvalds
        type: string
    type: object
  service.CelebrityData:
    properties:
      birth_date:
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - birth_date
    - name
    type: object
  service.Genre:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: horror
        type: string
    type: object
...
  service.Movie:
    properties:
      duration_minutes:
        example: 255
        type: integer
      id:
        example: 1
        type: integer
      poster_url:
        example: https://www.amazon.nl/Lord-Strings-J-Khan/dp/B000BU9TAQ
        type: string
      released_at:
        example: "2025-10-30T00:00:00Z"
        type: string
      synopsis:
        example: A man that can manipulate every string variable
        type: string
      title:
        example: Lord of the Strings
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esta es la API REST para la aplicación de gestión de películas.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API de Películas UNICEN
  version: "1.0"
paths:
  /celebrities:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de celebridades con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de celebridads exitosa
          schema:
            items:
              $ref: '#/definitions/service.Celebrity'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista celebridades paginadas
      tags:
      - Celebrities
    post:
      consumes:
      - application/json
      description: Agrega una nueva celebridad al catálogo.
      parameters:
      - description: Datos de la celebridad
        in: body
        name: celebrity
        required: true
        schema:
          $ref: '#/definitions/service.CelebrityData'
      produces:
      - application/json
      responses:
        "201":
          description: Celebridad agregada exitosamente
          schema:
            $ref: '#/definitions/service.Celebrity'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicacion de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega una celebridad
      tags:
      - Celebrities
  /celebrities/{ID}:
    delete:
      consumes:
      - application/json
      description: Borra una celebridad dada por ID permanentemente de la base de
        datos.
      parameters:
      - description: ID de la celebridad a eliminar
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Celebridad eliminada exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por celebridad no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra una celebridad
      tags:
      - Celebrities
    get:
      consumes:
      - application/json
      description: Obtiene toda la información asociada a una celebridad dada por
        ID.
      parameters:
      - description: ID de la celebridad a obtener
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Celebridad obtenida exitosamente
          schema:
            $ref: '#/definitions/service.Celebrity'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por celebridad no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Obtiene una celebridad
      tags:
      - Celebrities
    put:
      consumes:
      - application/json
      description: Actualiza toda la información de una celebridad dado por ID. Todos
        los campos son requeridos.
      parameters:
      - description: ID de la celebridad a actualizar
        in: path
        name: ID
        required: true
        type: integer
      - description: Datos actualizables de la celebridad
        in: body
        name: celebrity
        required: true
        schema:
          $ref: '#/definitions/service.CelebrityData'
      produces:
      - application/json
      responses:
        "200":
          description: Celebridad actualizada exitosamente
          schema:
            $ref: '#/definitions/service.Celebrity'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por celebridad no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Actualiza los datos de una celebridad
      tags:
      - Celebrities
...
  /genres/{genreID}/categories:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de películas asociadas a un género dado por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID del género
        in: path
        name: genreID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de películas exitosa
          schema:
            items:
              $ref: '#/definitions/service.Movie'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista películas de un género paginadas
      tags:
      - Categories
...
  /movies/{movieID}/categories:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de géneros asociados a una película dada por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de géneros exitosa
          schema:
            items:
              $ref: '#/definitions/service.Genre'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista géneros de una película paginados
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Agrega un género dado por ID a una película también dada por ID.
      parameters:
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      - description: Datos del género
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/service.CategoryData'
      produces:
      - application/json
      responses:
        "204":
          description: Género agregado exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error de película o género no encontrado
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega un género a una película
      tags:
      - Categories
  /movies/{movieID}/categories/{genreID}:
    delete:
      consumes:
      - application/json
      description: Borra un género dado por ID asociado a una película dada también
        por ID permanentemente.
      parameters:
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      - description: ID del género
        in: path
        name: genreID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Género borrado exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error de relación inexistente entre la película y el género
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra un género de una película
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Verifica si una película dada por ID tiene un determinado género
        también dado por ID.
      parameters:
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      - description: ID del género
        in: path
        name: genreID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Existe la relación entre la película y el género
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error de relación inexistente entre la película y el género
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Verifica que la película tiene asociado un género
      tags:
      - Categories
...
swagger: "2.0"